requirement.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.7
python-dotenv>=1.0.0
requests>=2.28.2
pyyaml>=6.0
xmltodict>=0.13.0
markdown>=3.4.3

# AI reporting - Basic (no advanced AI needed)
# None required - simple_ai_reporter.py works with standard library

# LangChain AI reporting (optional)
# langchain>=0.0.267
# langchain-community>=0.0.13
# openai>=0.27.0

# Alternative LLM providers (optional)
# anthropic>=0.2.0
# langchain-anthropic>=0.0.5
# ollama>=0.1.5
